#62 [EMS] The EMS heartbeat topic should be created by EMS service not EDS
