+++ docker-compose.yaml
@@ -10,6 +10,8 @@ services:
     environment:
       - EMS_KAFKA_BOOTSTRAP_SERVERS=etl-kafka:9092
       - EMS_KAFKA_TOPICS_HEARTBEAT=ems.heartbeat
+      - EMS_KAFKA_TOPICS_CONTROL=ems.control
+      - EMS_KAFKA_TOPICS_PROGRESS=ems.progress
       - SPRING_DATASOURCE_URL=jdbc:postgresql://etl-postgres-db:5432/etl
       - SPRING_DATASOURCE_USERNAME=etl
       - SPRING_DATASOURCE_PASSWORD=etl
