+++ extract-data-service/src/main/java/etl/engine/extract/config/KafkaConfiguration.java
@@ -2,13 +2,11 @@ package etl.engine.extract.config;
 
 import etl.engine.extract.service.messaging.CustomJsonSerializer;
 import org.apache.kafka.clients.admin.AdminClientConfig;
-import org.apache.kafka.clients.admin.NewTopic;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.common.serialization.StringSerializer;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.config.TopicBuilder;
 import org.springframework.kafka.core.DefaultKafkaProducerFactory;
 import org.springframework.kafka.core.KafkaAdmin;
 import org.springframework.kafka.core.KafkaTemplate;
@@ -22,9 +20,6 @@ public class KafkaConfiguration {
 
     @Value("${eds.kafka.bootstrap-servers}")
     private String bootstrapServers;
-
-    @Value("${eds.kafka.topics.heartbeat}")
-    private String heartBeatTopicName;
 
     @Bean
     public Map<String, Object> producerConfiguration() {
@@ -57,12 +52,5 @@ public class KafkaConfiguration {
         return new KafkaTemplate<>(producerFactory());
     }
 
-    @Bean
-    public NewTopic heartBeatTopic() {
-        //TODO What about retention policy?
-        return TopicBuilder
-                .name(heartBeatTopicName)
-                .partitions(1)
-                .build();
-    }
+
 }
